KuttaKoin - Initial Commit
KuttaKoin (KUTTA) is a unique ERC20 meme cryptocurrency inspired by a
popular dog meme, tailored to resonate with the vibrant community of
Indian crypto enthusiasts.

KuttaKoin aims to inject a sense of humor and camaraderie into the
Indian crypto community by leveraging the popularity of a dog meme as
its central theme. The cryptocurrency operates on the Ethereum
blockchain as an ERC20 token, ensuring compatibility with various
decentralized applications and wallets.